Dual Box Protection Improved! - Add new class for Dual Box Protection (singleton). - Translate all comments. - No more static.
